# [Fix](bangc-ops): optimize indice_convolution_forward